Fix exceptions
Exceptions now correctly show the stack trace.

This also went for me to fix "ifeq" opcode, which led me discover that for some weird reasons, intern() took way too long, making the test go to 150ms. And when return is commented the test goes down to 50ms. I think there is some latency with returning from a function that i'll have to check.

For now due to the fact that intern is by far the most called function (excluding constructors), it will probably be made into native code.